requirements.txt
flask
PyPDF2
pillow
reportlab